More robust code for special footnotes that have no fnmark
This should address some concerns raised before. Now we explicity check if @thefnmark is *empty* or not. If it's empty (such as for special footnotes like documentclass amsart's date) then only the footnote text is displayed, if it's *not* empty, we print the fnmark followed by a dot and a space (for French style footnotes).